+++ tests/testthat/test-labelled.R
@@ -33,3 +33,8 @@ test_that("zap_labels returns variables not of class('labelled') unmodified", {
   var <- c(1L, 98L, 99L)
   expect_equal(zap_labels(var), var)
 })
+
+test_that("zap_labels is applied to every column in data frame", {
+  df <- tibble::data_frame(x = 1:10, y = labelled(10:1, c("good" = 1)))
+  expect_equal(zap_labels(df)$y, 10:1)
+})